# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation de l'API pour la gestion des utilisateurs et l'authentification.",
        "contact": {
            "email": "sahliabderazak530@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/temoiniage": {
            "post": {
                "tags": [
                    "Témoignage"
                ],
                "summary": "Ajouter un témoignage",
                "description": "Permet à un utilisateur d'ajouter un témoignage, en attente de validation par un administrateur.",
                "operationId": "storeTestimonial",
                "requestBody": [
                    {
                        "mediaType": "application/json",
                        "schema": {
                            "required": [
                                "nom",
                                "email",
                                "temoignage"
                            ],
                            "properties": {
                                "nom": {
                                    "description": "Nom de l'utilisateur",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Email de l'utilisateur",
                                    "type": "string"
                                },
                                "temoignage": {
                                    "description": "Le contenu du témoignage",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Témoignage ajouté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "temoignage": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nom": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "temoignage": {
                                                    "type": "string"
                                                },
                                                "valider": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation des données"
                    }
                }
            }
        },
        "/api/temoignagesValides": {
            "get": {
                "tags": [
                    "Témoignage"
                ],
                "summary": "Récupérer tous les témoignages validés",
                "description": "Cette fonction récupère tous les témoignages validés par un administrateur.",
                "operationId": "getValidTestimonials",
                "responses": {
                    "200": {
                        "description": "Liste des témoignages validés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "temoignages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nom": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "temoignage": {
                                                        "type": "string"
                                                    },
                                                    "valider": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun témoignage validé trouvé"
                    }
                }
            }
        },
        "/api/temoiniages_admin": {
            "get": {
                "tags": [
                    "Témoignage"
                ],
                "summary": "Récupérer tous les témoignages",
                "description": "Permet à un utilisateur authentifié de récupérer tous les témoignages.",
                "operationId": "getAllTestimonials",
                "responses": {
                    "200": {
                        "description": "Liste de tous les témoignages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nom": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "temoignage": {
                                                "type": "string"
                                            },
                                            "valider": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    },
                    "404": {
                        "description": "Aucun témoignage trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/temoiniages/valider/{id}": {
            "put": {
                "tags": [
                    "Témoignage"
                ],
                "summary": "Valider un témoignage",
                "description": "Permet à un utilisateur authentifié de valider un témoignage.",
                "operationId": "validateTestimonial",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du témoignage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Témoignage validé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "temoiniage": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nom": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "temoignage": {
                                                    "type": "string"
                                                },
                                                "valider": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    },
                    "404": {
                        "description": "Témoignage introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/temoignageSupp/{id}": {
            "delete": {
                "tags": [
                    "Témoignage"
                ],
                "summary": "Supprimer un témoignage",
                "description": "Permet de supprimer un témoignage avec l'ID spécifié.",
                "operationId": "deleteTestimonial",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du témoignage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Témoignage supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Témoignage non trouvé"
                    }
                }
            }
        },
...
    },
...
}